refactor: fix Workspace component naming typos
