+++ src/components/BrowserAgentWorkspace/index.tsx
