+++ src/components/TerminalAgentWorkspace/index.tsx
